replace print with logger warning for invalid state
when the logit processor is finished, beams should never end up in an invalid state. This will log a warning for now. Experimental practice will show if this should be hardened to an assert or if there's some alternative that can force the beam to be abandoned without bringing down the whole thing