add a more sophisticated prompt for c1663
Adds context in the form of the line immedietly preceeding the anagram. Empty prompts and instructive prompts and structural context (i.e. write a line from a webcomic) are all leading to very strange results. If context from the comic is helpful, it's possible to go back and recalculate context for every line in the /data folder for additional testing