its the return of the heap
this is a thin passthrough to the heapq methods that take a heap as the first argument, wrapped in a class where the heap is managed by self.

It is missing all the underbar methods that make it practical to directly access the underlying list. Consider that in the future