add sophistication to scoring
min-heap means we want better scores the be smaller, but the oracle expresses better as higher scores. This flips that score so that it works with the heap.

this also adds a number of constraints specific to c1663. They only modify the score when violated, where they set the score to the worst possible value in order to fall out of the working queue quickly