candidate evaluation should short circuit
candidates that cannot be valid solutions due to words already placed should be abandoned quickly without recording.

This will mean we're not truly keeping a record of all visited nodes in the search tree, but it will also save us evaluating (and recording!) ~15k child nodes per-candidate. For c1663, a full third of the vocabulary is illegal from the start so this ends up being a lot of wasted space