add PersistentSearchQueue backed by SQLite
This largely replicates the interface of SearchQueue, but it cannot accept an existing iterable. The intent here is to be able to re-use the same SQLite DB between runs of the process, reusing the last PersistentQueue. Practically the search program initializes SearchQueues empty and its easier to disallow this than it is to figure out how to reconcile potentially pre-existing DB entries with a provided iterable.