perform random sampling fro PersistentSearchQueue in python
SQLite's non-standard random() and lack of log make the usual weighted-random-sampling-in-sql statements tricky. This eats the serialization cost of returning the whole frontier-set and then using python's random.choices to select the next node to explore. This is expensive, but the sampling function occurs infrequently and the whole table holds a maximum of 1000 rows. A profiler run can determine if its really a problem later.