overhaul candidate scoring
this patch reverts to using the log value returned from the oracle, and then multiplies it by the number of words in the sentence to create a candidate score.

The oracle's scoring mechanism favors fewer words because it is aggregating (faux) probabilities, and the odds of A -> B happening are a precondition of (and so necessarily higher) than (A -> B) -> C. In order to encourage exploration into the deeper solution space the score needs to incorporate other terms. This change incorporates a term to increase the scores for candidates that have used more of the available letters.