factor a raw score calculation method out of score_candidates
this new method returns the list of conditional next-token scores. The existing method calls this, then aggregates those results into a single score. This method will be useful with depth-first expansion of the search tree so that we can take a random walk until soft_validation fails and then query the oracle one time to get scores for all intermediary steps.