start a new search method
drafting a uct-inspired search method. Since there is only 1 or at most a very small number of winners, this replaces the win-loss ratios of UCT with the Oracle's BEAM score. Once we reach an unscored node, perform some number of uniform-random depth first explorations ending in a win or soft_validation failure. Pick a number that effriciently uses the vram (we can pass ten or maybe a hundred in at a time) and then backpropogate the discovered scores to the graph and start over again from the root. This should give us a UCT-style unbalanced search tree of the space adapted for our weird need-in-a-haystack-search problem