add a cumulative score column to the table schema
score is the score of moving from parent to placed. it makes it difficult to sort candidates in useful absolute way. By summing all the individual scores along the path we get a number that is more useful as an abstract comparator between runs, while the search algorithm still relies on the scores at each step.