selection phase should choose from already explored nodes
prior to this commit the search_tree.get method was always returning (implictly) None, resulting in every run of the loop doing monte carlo from the root. This change updates selection to use the new get implementation. It also sets up a weight_offset for the biased sample of which node to explore -- the sum of the weights must be positive but scores are negative, so we need to calculate an offset based off the lowest score and boost it by 1 in case all the scores are equal. This can occur when all nodes are new and the default exploration value is used.