add status codes
rather than ignoring entries based on score, this adds a status code that is used to filter words. An invalidating code is set at the same time as a -inf score. This sets up future infrastructure for status code 7, indicating manual intervention to indicate a fruitless search.

Many entries in the sqlite database are obviously incorrect to a human but may not get a bad score from the oracle until several additional words have been added. While this would eventually play out fine, judicious human intervention can save a lot of computation. Use extreme caution when deteriming "natural-sounding, reasonably-grammatical dialogue" because it would be very easy to prune out the winning sector of the tree.