PST.contains must limit responses in python
SQLite is case insensitive and so the call from the database may return inappropriate results that must be filtered in python. Because of this limits on the # of response rows must be applied after the last result filter or an artifically low number of rows may be returned instead, possibly even None when entries actually exist.