simplify trim
since we've already matched rows[0][0] to placed and status, it is safe to use those directly rather than iterating over the rows to find them.

after processing a direct match, remove it from rows before proceeding with deletion. If there is a direct match it will always be the first row returned because rows are ordered by placed latters, so can be removed with a slice.

also adds sql to match the case of placed ending the string.