add context to orace calls when available
This extends the oracle to use context, if provided. In the case of 1663, this uses the rest of the panel as left padding. Right now this is the only case that will use padding.

If padding changes, scores will change. You should start with an empty DB whenever you change context.

This change immediately improves the score of "I", and should improve all second word scores. Without context the oracle is extremely forgiving of second words that are obviously wrong to a human judge