rewrite c1663 flag
instead of trying to get three states out of a boolean flag, this change refactors it into a single flag to suppress c1663 heuristics and rules. Instead the program will infer c1663 to be true when the letter pool matches, and this flag will allow overriding the use of additional rules and heuristics only for that comic.

considering a more generic alternative where we use the comic data to make an arbitrary number of these flags, but that seemed premature. It was a goal for this to be modular, and using other (shorter) panels is certainly useful for proving the technique and writing tests. It's also possible that it could simplify some of the program logic to have a module where we can retrieve any comic #.

I rejected the idea now because the up front scope creep would be a lot in the middle of another refactor