add a verify function for the persistentsearchtree
This function is intended to check whether all the entries in the database have the same letters. It is very expensive. Two fields must be retrieved from each row, converted to Counters, combined, and stored in a counter.

The function returns a boolean indicating whether there is exactly 1 Counter, along with the most significant entry, it's count, the total number of entry's and the total count.