add assessment method to solver
assessment covers the global score_fragment functionality and a bit of the backpropgation functionality. The Oracle API doesn't have a good way of going from `str -> List[str]`, like a terminal node -> all of the intermediary nodes. However its efficient and useful. Since we really don't use other oracles that may be an area to simplify things in the future. Transformer Oracles are super expensive (SLOW) to even bootstrap in tests, but because the other oracles are not representative this just ends up meaning there are no tests for this stuff. An area for future optimization