fixup remove backpropogation method from Solver
backpropogation has been replaced with a combination of assessment and inline updating. While the MCTS algorithm calls this step backpropogation the term is loaded within ML and may be misleading. I'm not sure if we should keep it or not, especially as this deviates from the MCTS method. Can it be backpropogation if there's no chain rule in sight?