solver should enter the root into the pst on the first run
if the root is not in the PST, put it there before searching. This should occur for empty trees (first search) and when there is a new candidate specified to restrict the search to a subtree that is not already in the tree.

this also lets us remove the empty tree detection branch in the selection method, since the tree can no longer be empty when that method is called