rewrite score_candidates
this inlines calc_scores into score_candidates, and changes score_candidates to return a tuple of (word, scores). This is sensible because we've removed the abstract notion of an Oracle, freeing us to change the API as it suits the TransformerOracle.

Rather than one candidate turning into multiple outputs, this changes the API to return the paired word-level scores, leaving it to consumers to merge them however they want. This flexibility is necessary for ad-hoc exploration