when checking scores, check for pos -2, not pos 3
candidates could really stand to be a dataclass rather an a tuple, because this change is opaque. Currently there are three different methods for scoring a candidate, but practically we use only the one that comes right before the Status code. This makes negative indexing (indexing from the right) quite useful for the looming migration down to a single score.

But again, this would be clearer as a dataclass