change solver.assessment to set -inf scores when candidates are out of valid words
now that the check subcommand exists it's possible to call Solver.assessment on non-terminal nodes. This violates the assumption that it will only be called to judge candidates that are out of valid words, so instead we need to calculate that directly and operate the conditional off that result. This is more expensive, and there may be a more efficient way to do this