disable test_solver.test_solve for being flaky
even with a fixed seed there's a lot of variance in this output. it takes a long time to run, and at best it shows us that the solver.solve method doesn't raise an exception (rather than demonstrating correctness). This comments it out to disable it in automatic test suites, but leaves it because there may be some value in a developer manually running it occasionally