add auto-letters override when running check
letters must be provided as an argument, which is space delimeted. When providing candidates to check from bash, it is painful to reduce a candidate to letters to make it work as an arg.

This change provides a new option to check that will ignore any provided letters argument and instantiate the Solver instead with exactly the letters in the candidate, and c1663 set to False