add warning output for check cli command when using auto-letters
when c1663 is false, which is always is with auto-letters enabled, there should never be a soft validation failure, should never be a -inf score. It is surprisingly common when I check the training data set of actual dinocomics, which suggests a bug.

This checks for the condition and outputs explicit debugging commands to reproduce the test case that failed. While this is redundant with running a single test case, it's useful when batch testing many samples from the training set