vocabulary is all lower case
This is imprecise, but in a way that does not impact c1663 much. A proper fix would require recalculating a case sensitive vocabulary list and then manually deriving a new disallow list for c1663 (though maybe not so hard, since we know the only upper case letter is "I". It's probably possible to re-calculate, and then add new words to both lists)